additional Aggregation charts and bug fixes

+++ MOE.Common/Business/Bins/Bin.cs
@@ -6,7 +6,7 @@ namespace MOE.Common.Business.Bins
     {
         public DateTime Start { get; set; }
         public DateTime End { get; set; }
-        public int Sum { get; set; } = 0;
+        public double Sum { get; set; } = 0;
         public double Average { get; set; } = 0;
     }
 }
